Adding tracker to flow

+++ main.py
@@ -29,6 +29,7 @@ def process_messages(messages: List[dict], sqs_client: boto3.client):
         body = json.loads(message['Body'])
         s3_key = body['Records'][0]['s3']['object']['key']
         s3_bucket = body['Records'][0]['s3']['bucket']['name']
+        # Handles the formating of the sqs message
         object_key = unquote_plus(s3_key)
 
         process_video_file(s3_bucket, object_key)
@@ -51,6 +52,7 @@ def process_video_file(s3_bucket: str, s3_key: str):
     Returns:
     None
     """
+    print(f"Processing {s3_key}.......")
     try:
         s3 = boto3.client('s3', region_name='us-east-1')
         local_filename = '/tmp/' + s3_key.split('/')[-1]
@@ -58,17 +60,17 @@ def process_video_file(s3_bucket: str, s3_key: str):
         s3.download_file(s3_bucket, s3_key, local_filename)
     except Exception as e:
         print(f"Error downloading video: {e}")
-    try:
+   
         # Call your video processing module here
         #results_text, gif1, gif2 = video_processing_module.process_video(local_filename)
         results_text, gif1, gif2 = 'test', None, None
         # Upload results back to another S3 bucket
         upload_results('ag-video-results', s3_key, results_text, gif1, gif2)
-    except Exception as e:
-            print(f"Error uploading video: {e}")
+    
     finally:
         # Clean up: Delete the local video file
-        os.remove(local_filename)
+        if os.path.exists(local_filename):
+            os.remove(local_filename)
         # Clean up: Delete the local GIF files if they exist
         if os.path.exists(gif1):
             os.remove(gif1)
@@ -89,15 +91,18 @@ def upload_results(bucket_name: str, base_key: str, results_text: str, gif1: str
     Returns:
     None
     """
+    print(f"Uploading {results_text,gif1,gif2}.......")
+
     s3 = boto3.client('s3', region_name='us-east-1')
     result_prefix = base_key.replace('.mp4', '')  
-    
-    # Write text results
-    s3.put_object(Bucket=bucket_name, Key=result_prefix + '_results.txt', Body=results_text)
-    # Write gif results
-    #s3.upload_file(gif1, bucket_name, result_prefix + '_1.gif')
-    #3.upload_file(gif2, bucket_name, result_prefix + '_2.gif')
-
+    try:
+        # Write text results
+        s3.put_object(Bucket=bucket_name, Key=result_prefix + '_results.txt', Body=results_text)
+        # Write gif results
+        #s3.upload_file(gif1, bucket_name, result_prefix + '_1.gif')
+        #3.upload_file(gif2, bucket_name, result_prefix + '_2.gif')
+    except Exception as e:
+            print(f"Error uploading video: {e}")
 if __name__ == '__main__':
     # The URL of the SQS queue from which messages are received
     # Create SQS client
@@ -110,6 +115,7 @@ if __name__ == '__main__':
     while True:
         try:
             # Long polling for messages from the SQS queue to reduce the number of empty responses and lower costs
+            print("Polling SQS.......")
             response = sqs.receive_message(
                 QueueUrl=sqs_queue_url,
                 MaxNumberOfMessages=10,  # Retrieve up to 10 messages in one request
